Add section event detail to thank our partner for this event.
Add new picture and a text section in the event detail. We can add
the picture of the presenter and a little description of the personne
and a thankyou.

The top of the page is verticaly separated. On is right side, there
is the new section with the picture and the text.